some benchmarking stuff, will probably be removed again

+++ core/vtk/ttkMergeTreeDistanceMatrix/ttkMergeTreeDistanceMatrix.cpp
@@ -219,7 +219,7 @@ int ttkMergeTreeDistanceMatrix::run(
   execute<dataType>(intermediateTrees, intermediateTrees2, treesDistMat);
   std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
   auto time = std::chrono::duration_cast<std::chrono::milliseconds>(end - begin).count();
-  std::cout << time/1000 << "s" << std::endl;
+  std::cout << ((float)time)/1000. << "s" << std::endl;
 
   // --- Create output
   auto treesDistTable = vtkTable::GetData(outputVector);
